assaass

+++ clinicaOdontologica/src/main/java/com/digitalhouse/backend/integrador/clinicaOdontologica/dto/salida/odontologo/OdontologoSalidaDto.java
@@ -1,16 +1,16 @@
 package com.digitalhouse.backend.integrador.clinicaOdontologica.dto.salida.odontologo;
 
 public class OdontologoSalidaDto {
-    private int id;
+    private Long id;
     private int numeroMatricula;
     private String nombre;
     private String apellido;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -41,7 +41,7 @@ public class OdontologoSalidaDto {
     public OdontologoSalidaDto() {
     }
 
-    public OdontologoSalidaDto(int id, int numeroMatricula, String nombre, String apellido) {
+    public OdontologoSalidaDto(Long id, int numeroMatricula, String nombre, String apellido) {
         this.id = id;
         this.numeroMatricula = numeroMatricula;
         this.nombre = nombre;
