+++ clinicaOdontologica/src/main/java/com/digitalhouse/backend/integrador/clinicaOdontologica/dto/salida/paciente/DomicilioSalidaDto.java
@@ -1,17 +1,17 @@
 package com.digitalhouse.backend.integrador.clinicaOdontologica.dto.salida.paciente;
 
 public class DomicilioSalidaDto {
-    private int id;
+    private Long id;
     private String calle;
     private int numero;
     private String localidad;
     private String provincia;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -50,7 +50,7 @@ public class DomicilioSalidaDto {
     public DomicilioSalidaDto() {
     }
 
-    public DomicilioSalidaDto(int id, String calle, int numero, String localidad, String provincia) {
+    public DomicilioSalidaDto(Long id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
